requirements.txt
#bpy==3.4.0
mathutils==3.3.0

# linters
fake-bpy-module-3.4==20230117
black==22.12.0
pylint==2.15.9
isort==5.11.4

# pytest
pytest==7.2.1

# blender
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
Cython==0.29.26
idna==3.3
numpy==1.22.0
pycodestyle==2.8.0
requests==2.27.1
toml==0.10.2
urllib3==1.26.8
zstandard==0.16.0
